services/face-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
deepface==0.0.93
tensorflow-cpu==2.18.0   # CPU-only — no GPU needed for inference on Railway
tf-keras==2.18.0
retina-face==0.0.17
httpx==0.28.1
Pillow==11.1.0
numpy==1.26.4
pydantic==2.10.4
python-multipart==0.0.20
